refactor: método de autenticação refatorado e adicionado exceptions.
