+++ .env.example
@@ -1,2 +1,3 @@
 SP_TRANS_API_KEY=YOUR_API_KEY
 SP_TRANS_API_ENDPOINT=http://api.olhovivo.sptrans.com.br/
+SP_TRANS_API_VERSION=v2.1/
