+++ composer.json
@@ -13,16 +13,19 @@
         }
     ],
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "guzzlehttp/guzzle": "^7.5",
         "vlucas/phpdotenv": "^5.5"
     },
     "autoload": {
         "psr-4": {
-            "RafaelDuarte\\OlhoVivo\\": "src/"
+            "RafaelDuarte\\": "src/"
         }
     },
     "minimum-stability": "stable",
     "prefer-stable": true,
-    "version": "1.3.0"
+    "version": "1.3.0",
+    "require-dev": {
+        "symfony/var-dumper": "^6.2"
+    }
 }
