+++ src/Exceptions.php
@@ -1,0 +1,8 @@
+<?php
+namespace RafaelDuarte;
+
+use Exception;
+
+class Exceptions extends Exception
+{
+}
