+++ supervisord.conf
@@ -5,7 +5,7 @@ logfile=/var/log/supervisor/supervisord.log
 pidfile=/var/run/supervisord.pid
 
 [program:php]
-command=php -S 0.0.0.0:80 -t /var/www/html/src -d post_max_size=100M -d upload_max_filesize=100M -d variables_order=EGPCS -d opcache.enable_cli=1
+command=php -S 0.0.0.0:80 -t /var/www/html/src
 user=app
 stdout_logfile=/var/log/supervisor/php-serve.log
 stdout_logfile_maxbytes=0
